CAT-1360 Adjust ChangeColorByN and SetColorTo Bricks

+++ catroidTest/src/org/catrobat/catroid/test/formulaeditor/ParserTestObject.java
@@ -46,7 +46,7 @@ public class ParserTestObject extends AndroidTestCase {
 	private static final float LOOK_Y_POSITION = 23.4f;
 	private static final float LOOK_X_POSITION = 5.6f;
 	private static final float LOOK_BRIGHTNESS = 0.7f;
-	private static final int LOOK_COLOR = 0;
+	private static final float LOOK_COLOR = 0f;
 	private static final float LOOK_SCALE = 90.3f;
 	private static final float LOOK_ROTATION = 30.7f;
 	private static final float DELTA = 0.01f;
